requirements.txt
gTTS
playsound
twitch-python
